requirements.txt
Flask==3.0.3
mysql-connector-python==8.0.23
requests==2.25.1
beautifulsoup4==4.9.3
selenium==3.141.0
webdriver-manager==3.4.2
